solucao2/requirements.txt
fastapi
uvicorn
databases
passlib
python-jose
python-multipart
bcrypt
